Revert a change to fix installer window interaction
We found that this commit causes a bug where, sometimes, you
can't interact with windows in the installer's advanced
partitioning mode. Reverting this change means drop shadows are
again considered part of a window's target area for clicking,
but that seems like much less of a problem than "sometimes you
can't click on anything at all".

https://bugzilla.redhat.com/show_bug.cgi?id=2239128